Refactored package structure of services layer
Added /services with __init__ py.
Whole code from services.py moved to various modules in /services.
Implemented facade in __init__.py using __all__.
